+++ apis/dataproc/v1alpha1/types.generated.go
@@ -40,54 +40,6 @@ type AcceleratorConfig struct {
 	AcceleratorCount *int32 `json:"acceleratorCount,omitempty"`
 }
 
-// +kcc:proto=google.cloud.dataproc.v1.DiskConfig
-type DiskConfig struct {
-	// Optional. Type of the boot disk (default is "pd-standard").
-	//  Valid values: "pd-balanced" (Persistent Disk Balanced Solid State Drive),
-	//  "pd-ssd" (Persistent Disk Solid State Drive),
-	//  or "pd-standard" (Persistent Disk Hard Disk Drive).
-	//  See [Disk types](https://cloud.google.com/compute/docs/disks#disk-types).
-	// +kcc:proto:field=google.cloud.dataproc.v1.DiskConfig.boot_disk_type
-	BootDiskType *string `json:"bootDiskType,omitempty"`
-
-	// Optional. Size in GB of the boot disk (default is 500GB).
-	// +kcc:proto:field=google.cloud.dataproc.v1.DiskConfig.boot_disk_size_gb
-	BootDiskSizeGB *int32 `json:"bootDiskSizeGB,omitempty"`
-
-	// Optional. Number of attached SSDs, from 0 to 8 (default is 0).
-	//  If SSDs are not attached, the boot disk is used to store runtime logs and
-	//  [HDFS](https://hadoop.apache.org/docs/r1.2.1/hdfs_user_guide.html) data.
-	//  If one or more SSDs are attached, this runtime bulk
-	//  data is spread across them, and the boot disk contains only basic
-	//  config and installed binaries.
-	//
-	//  Note: Local SSD options may vary by machine type and number of vCPUs
-	//  selected.
-	// +kcc:proto:field=google.cloud.dataproc.v1.DiskConfig.num_local_ssds
-	NumLocalSsds *int32 `json:"numLocalSsds,omitempty"`
-
-	// Optional. Interface type of local SSDs (default is "scsi").
-	//  Valid values: "scsi" (Small Computer System Interface),
-	//  "nvme" (Non-Volatile Memory Express).
-	//  See [local SSD
-	//  performance](https://cloud.google.com/compute/docs/disks/local-ssd#performance).
-	// +kcc:proto:field=google.cloud.dataproc.v1.DiskConfig.local_ssd_interface
-	LocalSsdInterface *string `json:"localSsdInterface,omitempty"`
-
-	// Optional. Indicates how many IOPS to provision for the disk. This sets the
-	//  number of I/O operations per second that the disk can handle. Note: This
-	//  field is only supported if boot_disk_type is hyperdisk-balanced.
-	// +kcc:proto:field=google.cloud.dataproc.v1.DiskConfig.boot_disk_provisioned_iops
-	BootDiskProvisionedIops *int64 `json:"bootDiskProvisionedIops,omitempty"`
-
-	// Optional. Indicates how much throughput to provision for the disk. This
-	//  sets the number of throughput mb per second that the disk can handle.
-	//  Values must be greater than or equal to 1. Note: This field is only
-	//  supported if boot_disk_type is hyperdisk-balanced.
-	// +kcc:proto:field=google.cloud.dataproc.v1.DiskConfig.boot_disk_provisioned_throughput
-	BootDiskProvisionedThroughput *int64 `json:"bootDiskProvisionedThroughput,omitempty"`
-}
-
 // +kcc:proto=google.cloud.dataproc.v1.InstanceFlexibilityPolicy
 type InstanceFlexibilityPolicy struct {
 	// Optional. Defines how the Group selects the provisioning model to ensure
