+++ config/crds/resources/apiextensions.k8s.io_v1_customresourcedefinition_dataprocnodegroups.dataproc.cnrm.cloud.google.com.yaml
@@ -112,7 +112,7 @@ spec:
                   diskConfig:
                     description: Optional. Disk option config settings.
                     properties:
-                      bootDiskProvisionedIops:
+                      bootDiskProvisionedIOPs:
                         description: 'Optional. Indicates how many IOPS to provision
                           for the disk. This sets the number of I/O operations per
                           second that the disk can handle. Note: This field is only
@@ -139,13 +139,13 @@ spec:
                           Solid State Drive), or "pd-standard" (Persistent Disk Hard
                           Disk Drive). See [Disk types](https://cloud.google.com/compute/docs/disks#disk-types).'
                         type: string
-                      localSsdInterface:
+                      localSSDInterface:
                         description: 'Optional. Interface type of local SSDs (default
                           is "scsi"). Valid values: "scsi" (Small Computer System
                           Interface), "nvme" (Non-Volatile Memory Express). See [local
                           SSD performance](https://cloud.google.com/compute/docs/disks/local-ssd#performance).'
                         type: string
-                      numLocalSsds:
+                      numLocalSSDs:
                         description: |-
                           Optional. Number of attached SSDs, from 0 to 8 (default is 0).
                            If SSDs are not attached, the boot disk is used to store runtime logs and
